StudEsprit/requirements.txt
Django>=5.0,<6.0
pymongo>=4.6.0
python-dotenv>=1.0.1
argon2-cffi>=23.1.0
python-slugify>=8.0.4
ruff>=0.5.7
black>=24.8.0
requests>=2.31.0
google-auth>=2.34.0
